+++ integration_tests/gdrive/test_gdrive.py
@@ -2,27 +2,37 @@ import base64
 import json
 
 import pytest
+from google.oauth2.service_account import Credentials as ServiceCredentials
 
 import pathway as pw
+from pathway.io.gdrive import _ListObjectsStrategy
 
 TEST_FILE_SIZE = 1550383
 FOLDER_WITH_ONE_FILE_ID = "1XisWrSjKMCx2jfUW8OSgt6L8veq8c4Mh"
 FOLDER_WITH_SYMLINK_ID = "1wS3IdC1oNLxeTV5ZOqGEmz8z15kRmsBt"
-FOLDER_WITH_TYPES = "19B2X3HxfCsCh2NQ8-uzrvkuKUV2gqzdj"
+FOLDER_WITH_2047_FOLDERS = "1MI1qu49ZjOsmJC_CPlKmXur--fevOLnN"
+FOLDER_WITH_128_FOLDERS = "1oOfVCMjJrmKvxlun8_BqkdXJIwK_pMY5"
+
 # https://drive.google.com/drive/u/0/folders/19B2X3HxfCsCh2NQ8-uzrvkuKUV2gqzdj
+FOLDER_WITH_TYPES = "19B2X3HxfCsCh2NQ8-uzrvkuKUV2gqzdj"
 
 
+@pytest.mark.parametrize(
+    "list_objects_strategy",
+    [_ListObjectsStrategy.TreeTraversal, _ListObjectsStrategy.FullScan, None],
+)
 @pytest.mark.parametrize("object_size_limit", [None, 100, 2000000])
 @pytest.mark.parametrize("with_metadata", [False, True])
 def test_single_file_read_with_constraints(
-    object_size_limit, with_metadata, tmp_path, credentials_dir
+    list_objects_strategy, object_size_limit, with_metadata, tmp_path, credentials_dir
 ):
     files_table = pw.io.gdrive.read(
         FOLDER_WITH_ONE_FILE_ID,
         mode="static",
-        service_user_credentials_file=str(credentials_dir / "credentials.json"),
+        service_user_credentials_file=credentials_dir / "credentials.json",
         object_size_limit=object_size_limit,
         with_metadata=with_metadata,
+        _list_objects_strategy=list_objects_strategy,
     )
     pw.io.jsonlines.write(files_table, tmp_path / "output.jsonl")
     pw.run()
@@ -44,15 +54,22 @@ def test_single_file_read_with_constraints(
     assert rows_count == 1
 
 
+@pytest.mark.parametrize(
+    "list_objects_strategy",
+    [_ListObjectsStrategy.TreeTraversal, _ListObjectsStrategy.FullScan, None],
+)
 @pytest.mark.parametrize("object_size_limit", [None, 100, 2000000])
 @pytest.mark.parametrize("with_metadata", [True, False])
-def test_gdrive_symlink(object_size_limit, with_metadata, tmp_path, credentials_dir):
+def test_gdrive_symlink(
+    list_objects_strategy, object_size_limit, with_metadata, tmp_path, credentials_dir
+):
     files_table = pw.io.gdrive.read(
         FOLDER_WITH_SYMLINK_ID,
         mode="static",
-        service_user_credentials_file=str(credentials_dir / "credentials.json"),
+        service_user_credentials_file=credentials_dir / "credentials.json",
         object_size_limit=object_size_limit,
         with_metadata=with_metadata,
+        _list_objects_strategy=list_objects_strategy,
     )
     pw.io.jsonlines.write(files_table, tmp_path / "output.jsonl")
     pw.run()
@@ -68,6 +85,10 @@ def test_gdrive_symlink(object_size_limit, with_metadata, tmp_path, credentials_
     assert rows_count == 1
 
 
+@pytest.mark.parametrize(
+    "list_objects_strategy",
+    [_ListObjectsStrategy.TreeTraversal, _ListObjectsStrategy.FullScan, None],
+)
 @pytest.mark.parametrize("format", ["binary", "only_metadata"])
 @pytest.mark.parametrize("with_metadata", [False, True])
 @pytest.mark.parametrize(
@@ -84,7 +105,12 @@ def test_gdrive_symlink(object_size_limit, with_metadata, tmp_path, credentials_
     ],
 )
 def test_name_pattern_single_filter(
-    format, with_metadata, name_pattern, tmp_path, credentials_dir
+    list_objects_strategy,
+    format,
+    with_metadata,
+    name_pattern,
+    tmp_path,
+    credentials_dir,
 ):
     object_size_limit = None
 
@@ -97,10 +123,11 @@ def test_name_pattern_single_filter(
         FOLDER_WITH_TYPES,
         mode="static",
         format=format,
-        service_user_credentials_file=str(credentials_dir / "credentials.json"),
+        service_user_credentials_file=credentials_dir / "credentials.json",
         object_size_limit=object_size_limit,
         with_metadata=with_metadata,
         file_name_pattern=name_pattern,
+        _list_objects_strategy=list_objects_strategy,
     )
 
     pw.io.jsonlines.write(files_table, tmp_path / "output.jsonl")
@@ -140,3 +167,54 @@ def test_name_pattern_single_filter(
         assert (
             rows_count == NUM_TXT_FILES + NUM_CSV_FILES + NUM_MD_FILES + NUM_PDF_FILES
         )
+
+
+def test_gdrive_tree_traversal(credentials_dir, tmp_path):
+    files_table = pw.io.gdrive.read(
+        FOLDER_WITH_128_FOLDERS,
+        mode="static",
+        service_user_credentials_file=credentials_dir / "credentials.json",
+        format="only_metadata",
+    )
+    pw.io.jsonlines.write(files_table, tmp_path / "output.jsonl")
+    pw.run()
+    rows_count = 0
+    with open(tmp_path / "output.jsonl", "r") as f:
+        for _ in f:
+            rows_count += 1
+    assert rows_count == 129  # 128 files + python script that created them
+
+    credentials = ServiceCredentials.from_service_account_file(
+        credentials_dir / "credentials.json"
+    )
+    client = pw.io.gdrive._GDriveClient(
+        root=FOLDER_WITH_128_FOLDERS,
+        credentials=credentials,
+    )
+    assert len(client._traverse_objects_with_limit()) == 129
+
+
+def test_gdrive_full_scan(credentials_dir, tmp_path):
+    files_table = pw.io.gdrive.read(
+        FOLDER_WITH_2047_FOLDERS,
+        mode="static",
+        service_user_credentials_file=credentials_dir / "credentials.json",
+        format="only_metadata",
+    )
+    pw.io.jsonlines.write(files_table, tmp_path / "output.jsonl")
+    pw.run()
+    rows_count = 0
+    with open(tmp_path / "output.jsonl", "r") as f:
+        for _ in f:
+            rows_count += 1
+    assert rows_count == 2048  # 2047 files + python script that created them
+
+    credentials = ServiceCredentials.from_service_account_file(
+        credentials_dir / "credentials.json"
+    )
+    client = pw.io.gdrive._GDriveClient(
+        root=FOLDER_WITH_2047_FOLDERS,
+        credentials=credentials,
+    )
+    with pytest.raises(pw.io.gdrive._ListObjectsLimitExceeded):
+        client._traverse_objects_with_limit()